kinda add event system? (TODO: make a way to generally handle events)

+++ crates/bizarre_core/src/app.rs
@@ -4,7 +4,7 @@ use bizarre_logger::{core_debug, core_info, info};
 use bizarre_render::renderer::{create_renderer, Renderer, RendererBackend};
 use winit::platform::run_return::EventLoopExtRunReturn;
 
-use crate::event::key_codes::KeyboardKey;
+use crate::input::key_codes::KeyboardKey;
 
 pub struct AppConfig {
     pub title: Box<str>,
@@ -83,8 +83,8 @@ impl App {
                         *control_flow = winit::event_loop::ControlFlow::Exit;
                     }
                     winit::event::WindowEvent::KeyboardInput { input, .. } => {
-                        if (input.state == winit::event::ElementState::Pressed) {
-                            let keycode = KeyboardKey::from(input.scancode);
+                        if input.state == winit::event::ElementState::Pressed {
+                            let keycode = KeyboardKey::from(input.scancode as u16);
                             core_debug!("Keyboard input: {}", keycode);
                         }
                     }
